no quickmark_config
