+++ crates/quickmark/Cargo.toml
@@ -10,7 +10,6 @@ path = "src/main.rs"
 [dependencies]
 anyhow = "1.0.86"
 clap = { version = "4.5.4", features = ["derive"] }
-quickmark_config = { path = "../quickmark_config" }
 quickmark_linter = { path = "../quickmark_linter" }
 
 [dev-dependencies.quickmark_linter]
