--- crates/quickmark_config/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "quickmark_config"
-version = "0.0.1"
-edition = "2021"
-
-[dependencies]
-anyhow = "1.0.86"
-quickmark_linter = { path = "../quickmark_linter" }
-toml = "0.8.14"
-
-[dev-dependencies]
-tempfile = "3.8"
