+++ crates/quickmark_linter/Cargo.toml
@@ -9,8 +9,12 @@ linkify = "0.10"
 once_cell = "1.19"
 regex = "1.0"
 serde = { version = "1.0", features = ["derive"] }
+toml = "0.8.14"
 tree-sitter = "0.25.6"
 tree-sitter-md = "0.3.2"
 
 [features]
 testing = []
+
+[dev-dependencies]
+tempfile = "3.8"
