+++ crates/quickmark_linter/src/config/mod.rs
@@ -1,5 +1,7 @@
 use std::collections::{HashMap, HashSet};
+use std::{fs, path::Path};
 use serde::Deserialize;
+use anyhow::Result;
 
 use crate::rules::ALL_RULES;
 
@@ -606,6 +608,331 @@ impl QuickmarkConfig {
         normalize_severities(&mut config.linters.severity);
         config
     }
+}
+
+// TOML conversion functions
+fn convert_toml_severity(toml_severity: TomlRuleSeverity) -> RuleSeverity {
+    match toml_severity {
+        TomlRuleSeverity::Error => RuleSeverity::Error,
+        TomlRuleSeverity::Warning => RuleSeverity::Warning,
+        TomlRuleSeverity::Off => RuleSeverity::Off,
+    }
+}
+
+fn convert_toml_heading_style(toml_style: TomlHeadingStyle) -> HeadingStyle {
+    match toml_style {
+        TomlHeadingStyle::Consistent => HeadingStyle::Consistent,
+        TomlHeadingStyle::Atx => HeadingStyle::ATX,
+        TomlHeadingStyle::Setext => HeadingStyle::Setext,
+        TomlHeadingStyle::ATXClosed => HeadingStyle::ATXClosed,
+        TomlHeadingStyle::SetextWithATX => HeadingStyle::SetextWithATX,
+        TomlHeadingStyle::SetextWithATXClosed => HeadingStyle::SetextWithATXClosed,
+    }
+}
+
+fn convert_toml_ul_style(toml_style: TomlUlStyle) -> UlStyle {
+    match toml_style {
+        TomlUlStyle::Asterisk => UlStyle::Asterisk,
+        TomlUlStyle::Consistent => UlStyle::Consistent,
+        TomlUlStyle::Dash => UlStyle::Dash,
+        TomlUlStyle::Plus => UlStyle::Plus,
+        TomlUlStyle::Sublist => UlStyle::Sublist,
+    }
+}
+
+fn convert_toml_ol_prefix_style(toml_style: TomlOlPrefixStyle) -> OlPrefixStyle {
+    match toml_style {
+        TomlOlPrefixStyle::One => OlPrefixStyle::One,
+        TomlOlPrefixStyle::Ordered => OlPrefixStyle::Ordered,
+        TomlOlPrefixStyle::OneOrOrdered => OlPrefixStyle::OneOrOrdered,
+        TomlOlPrefixStyle::Zero => OlPrefixStyle::Zero,
+    }
+}
+
+fn convert_toml_code_block_style(toml_style: TomlCodeBlockStyle) -> CodeBlockStyle {
+    match toml_style {
+        TomlCodeBlockStyle::Consistent => CodeBlockStyle::Consistent,
+        TomlCodeBlockStyle::Fenced => CodeBlockStyle::Fenced,
+        TomlCodeBlockStyle::Indented => CodeBlockStyle::Indented,
+    }
+}
+
+fn convert_toml_code_fence_style(toml_style: TomlCodeFenceStyle) -> CodeFenceStyle {
+    match toml_style {
+        TomlCodeFenceStyle::Consistent => CodeFenceStyle::Consistent,
+        TomlCodeFenceStyle::Backtick => CodeFenceStyle::Backtick,
+        TomlCodeFenceStyle::Tilde => CodeFenceStyle::Tilde,
+    }
+}
+
+fn convert_toml_emphasis_style(toml_style: TomlEmphasisStyle) -> EmphasisStyle {
+    match toml_style {
+        TomlEmphasisStyle::Consistent => EmphasisStyle::Consistent,
+        TomlEmphasisStyle::Asterisk => EmphasisStyle::Asterisk,
+        TomlEmphasisStyle::Underscore => EmphasisStyle::Underscore,
+    }
+}
+
+fn convert_toml_strong_style(toml_style: TomlStrongStyle) -> StrongStyle {
+    match toml_style {
+        TomlStrongStyle::Consistent => StrongStyle::Consistent,
+        TomlStrongStyle::Asterisk => StrongStyle::Asterisk,
+        TomlStrongStyle::Underscore => StrongStyle::Underscore,
+    }
+}
+
+fn convert_toml_table_pipe_style(toml_style: TomlTablePipeStyle) -> TablePipeStyle {
+    match toml_style {
+        TomlTablePipeStyle::Consistent => TablePipeStyle::Consistent,
+        TomlTablePipeStyle::LeadingAndTrailing => TablePipeStyle::LeadingAndTrailing,
+        TomlTablePipeStyle::LeadingOnly => TablePipeStyle::LeadingOnly,
+        TomlTablePipeStyle::TrailingOnly => TablePipeStyle::TrailingOnly,
+        TomlTablePipeStyle::NoLeadingOrTrailing => TablePipeStyle::NoLeadingOrTrailing,
+    }
+}
+
+/// Parse a TOML configuration string into a QuickmarkConfig
+pub fn parse_toml_config(config_str: &str) -> Result<QuickmarkConfig> {
+    let toml_config: TomlQuickmarkConfig = toml::from_str(config_str)?;
+    let mut severity = toml_config
+        .linters
+        .severity
+        .into_iter()
+        .map(|(k, v)| (k, convert_toml_severity(v)))
+        .collect();
+
+    normalize_severities(&mut severity);
+
+    Ok(QuickmarkConfig::new(LintersTable {
+        severity,
+        settings: LintersSettingsTable {
+            heading_style: MD003HeadingStyleTable {
+                style: convert_toml_heading_style(toml_config.linters.settings.heading_style.style),
+            },
+            ul_style: MD004UlStyleTable {
+                style: convert_toml_ul_style(toml_config.linters.settings.ul_style.style),
+            },
+            ol_prefix: MD029OlPrefixTable {
+                style: convert_toml_ol_prefix_style(toml_config.linters.settings.ol_prefix.style),
+            },
+            ul_indent: MD007UlIndentTable {
+                indent: toml_config.linters.settings.ul_indent.indent,
+                start_indent: toml_config.linters.settings.ul_indent.start_indent,
+                start_indented: toml_config.linters.settings.ul_indent.start_indented,
+            },
+            trailing_spaces: MD009TrailingSpacesTable {
+                br_spaces: toml_config.linters.settings.trailing_spaces.br_spaces,
+                list_item_empty_lines: toml_config
+                    .linters
+                    .settings
+                    .trailing_spaces
+                    .list_item_empty_lines,
+                strict: toml_config.linters.settings.trailing_spaces.strict,
+            },
+            hard_tabs: MD010HardTabsTable {
+                code_blocks: toml_config.linters.settings.hard_tabs.code_blocks,
+                ignore_code_languages: toml_config.linters.settings.hard_tabs.ignore_code_languages,
+                spaces_per_tab: toml_config.linters.settings.hard_tabs.spaces_per_tab,
+            },
+            multiple_blank_lines: MD012MultipleBlankLinesTable {
+                maximum: toml_config.linters.settings.multiple_blank_lines.maximum,
+            },
+            line_length: MD013LineLengthTable {
+                line_length: toml_config.linters.settings.line_length.line_length,
+                code_block_line_length: toml_config
+                    .linters
+                    .settings
+                    .line_length
+                    .code_block_line_length,
+                heading_line_length: toml_config.linters.settings.line_length.heading_line_length,
+                code_blocks: toml_config.linters.settings.line_length.code_blocks,
+                headings: toml_config.linters.settings.line_length.headings,
+                tables: toml_config.linters.settings.line_length.tables,
+                strict: toml_config.linters.settings.line_length.strict,
+                stern: toml_config.linters.settings.line_length.stern,
+            },
+            headings_blanks: MD022HeadingsBlanksTable {
+                lines_above: toml_config.linters.settings.headings_blanks.lines_above,
+                lines_below: toml_config.linters.settings.headings_blanks.lines_below,
+            },
+            single_h1: MD025SingleH1Table {
+                level: toml_config.linters.settings.single_h1.level,
+                front_matter_title: toml_config.linters.settings.single_h1.front_matter_title,
+            },
+            first_line_heading: MD041FirstLineHeadingTable {
+                allow_preamble: toml_config
+                    .linters
+                    .settings
+                    .first_line_heading
+                    .allow_preamble,
+                front_matter_title: toml_config
+                    .linters
+                    .settings
+                    .first_line_heading
+                    .front_matter_title,
+                level: toml_config.linters.settings.first_line_heading.level,
+            },
+            trailing_punctuation: MD026TrailingPunctuationTable {
+                punctuation: toml_config
+                    .linters
+                    .settings
+                    .trailing_punctuation
+                    .punctuation,
+            },
+            blockquote_spaces: MD027BlockquoteSpacesTable {
+                list_items: toml_config.linters.settings.blockquote_spaces.list_items,
+            },
+            list_marker_space: MD030ListMarkerSpaceTable {
+                ul_single: toml_config.linters.settings.list_marker_space.ul_single,
+                ol_single: toml_config.linters.settings.list_marker_space.ol_single,
+                ul_multi: toml_config.linters.settings.list_marker_space.ul_multi,
+                ol_multi: toml_config.linters.settings.list_marker_space.ol_multi,
+            },
+            fenced_code_blanks: MD031FencedCodeBlanksTable {
+                list_items: toml_config.linters.settings.fenced_code_blanks.list_items,
+            },
+            inline_html: MD033InlineHtmlTable {
+                allowed_elements: toml_config.linters.settings.inline_html.allowed_elements,
+            },
+            hr_style: MD035HrStyleTable {
+                style: toml_config.linters.settings.hr_style.style,
+            },
+            emphasis_as_heading: MD036EmphasisAsHeadingTable {
+                punctuation: toml_config.linters.settings.emphasis_as_heading.punctuation,
+            },
+            fenced_code_language: MD040FencedCodeLanguageTable {
+                allowed_languages: toml_config
+                    .linters
+                    .settings
+                    .fenced_code_language
+                    .allowed_languages,
+                language_only: toml_config
+                    .linters
+                    .settings
+                    .fenced_code_language
+                    .language_only,
+            },
+            code_block_style: MD046CodeBlockStyleTable {
+                style: convert_toml_code_block_style(
+                    toml_config.linters.settings.code_block_style.style,
+                ),
+            },
+            code_fence_style: MD048CodeFenceStyleTable {
+                style: convert_toml_code_fence_style(
+                    toml_config.linters.settings.code_fence_style.style,
+                ),
+            },
+            emphasis_style: MD049EmphasisStyleTable {
+                style: convert_toml_emphasis_style(
+                    toml_config.linters.settings.emphasis_style.style,
+                ),
+            },
+            strong_style: MD050StrongStyleTable {
+                style: convert_toml_strong_style(toml_config.linters.settings.strong_style.style),
+            },
+            multiple_headings: MD024MultipleHeadingsTable {
+                siblings_only: toml_config.linters.settings.multiple_headings.siblings_only,
+                allow_different_nesting: toml_config
+                    .linters
+                    .settings
+                    .multiple_headings
+                    .allow_different_nesting,
+            },
+            link_fragments: MD051LinkFragmentsTable {
+                ignore_case: toml_config.linters.settings.link_fragments.ignore_case,
+                ignored_pattern: toml_config.linters.settings.link_fragments.ignored_pattern,
+            },
+            reference_links_images: MD052ReferenceLinksImagesTable {
+                shortcut_syntax: toml_config
+                    .linters
+                    .settings
+                    .reference_links_images
+                    .shortcut_syntax,
+                ignored_labels: toml_config
+                    .linters
+                    .settings
+                    .reference_links_images
+                    .ignored_labels,
+            },
+            required_headings: MD043RequiredHeadingsTable {
+                headings: toml_config.linters.settings.required_headings.headings,
+                match_case: toml_config.linters.settings.required_headings.match_case,
+            },
+            proper_names: MD044ProperNamesTable {
+                names: toml_config.linters.settings.proper_names.names,
+                code_blocks: toml_config.linters.settings.proper_names.code_blocks,
+                html_elements: toml_config.linters.settings.proper_names.html_elements,
+            },
+            link_image_reference_definitions: MD053LinkImageReferenceDefinitionsTable {
+                ignored_definitions: toml_config
+                    .linters
+                    .settings
+                    .link_image_reference_definitions
+                    .ignored_definitions,
+            },
+            link_image_style: MD054LinkImageStyleTable {
+                autolink: toml_config.linters.settings.link_image_style.autolink,
+                inline: toml_config.linters.settings.link_image_style.inline,
+                full: toml_config.linters.settings.link_image_style.full,
+                collapsed: toml_config.linters.settings.link_image_style.collapsed,
+                shortcut: toml_config.linters.settings.link_image_style.shortcut,
+                url_inline: toml_config.linters.settings.link_image_style.url_inline,
+            },
+            table_pipe_style: MD055TablePipeStyleTable {
+                style: convert_toml_table_pipe_style(
+                    toml_config.linters.settings.table_pipe_style.style,
+                ),
+            },
+            descriptive_link_text: MD059DescriptiveLinkTextTable {
+                prohibited_texts: toml_config
+                    .linters
+                    .settings
+                    .descriptive_link_text
+                    .prohibited_texts,
+            },
+        },
+    }))
+}
+
+/// Load configuration from QUICKMARK_CONFIG environment variable, path, or default
+pub fn config_from_env_path_or_default(path: &Path) -> Result<QuickmarkConfig> {
+    // First check if QUICKMARK_CONFIG environment variable is set
+    if let Ok(env_config_path) = std::env::var("QUICKMARK_CONFIG") {
+        let env_config_file = Path::new(&env_config_path);
+        if env_config_file.is_file() {
+            match fs::read_to_string(env_config_file) {
+                Ok(config) => return parse_toml_config(&config),
+                Err(e) => {
+                    eprintln!(
+                        "Error loading config from QUICKMARK_CONFIG path {env_config_path}: {e}. Default config will be used."
+                    );
+                    return Ok(QuickmarkConfig::default_with_normalized_severities());
+                }
+            }
+        } else {
+            eprintln!(
+                "Config file was not found at QUICKMARK_CONFIG path {env_config_path}. Default config will be used."
+            );
+            return Ok(QuickmarkConfig::default_with_normalized_severities());
+        }
+    }
+
+    // Fallback to existing behavior - check for quickmark.toml in path
+    config_in_path_or_default(path)
+}
+
+/// Load configuration from a path, or return default if not found
+pub fn config_in_path_or_default(path: &Path) -> Result<QuickmarkConfig> {
+    let config_file = path.join("quickmark.toml");
+    if config_file.is_file() {
+        let config = fs::read_to_string(config_file)?;
+        return parse_toml_config(&config);
+    }
+    println!(
+        "Config file was not found at {}. Default config will be used.",
+        config_file.to_string_lossy()
+    );
+    Ok(QuickmarkConfig::default_with_normalized_severities())
 }
 
 // TOML configuration structures for serialization/deserialization
@@ -1410,8 +1737,10 @@ impl Default for TomlMD055TablePipeStyleTable {
 #[cfg(test)]
 mod test {
     use std::collections::HashMap;
+    use std::path::Path;
 
     use crate::config::{
+        config_from_env_path_or_default, parse_toml_config,
         HeadingStyle, LintersSettingsTable, LintersTable, MD003HeadingStyleTable,
         MD004UlStyleTable, MD007UlIndentTable, MD009TrailingSpacesTable, MD010HardTabsTable,
         MD012MultipleBlankLinesTable, MD013LineLengthTable, MD022HeadingsBlanksTable,
@@ -1424,6 +1753,7 @@ mod test {
         MD051LinkFragmentsTable, MD052ReferenceLinksImagesTable,
         MD053LinkImageReferenceDefinitionsTable, MD054LinkImageStyleTable,
         MD055TablePipeStyleTable, MD059DescriptiveLinkTextTable, RuleSeverity,
+        UlStyle,
     };
 
     use super::{normalize_severities, QuickmarkConfig};
@@ -1538,6 +1868,168 @@ mod test {
         assert_eq!(
             HeadingStyle::ATX,
             config.linters.settings.heading_style.style
+        );
+    }
+
+    // TOML parsing tests
+    #[test]
+    fn test_parse_md028_config() {
+        let config_str = r#"
+        [linters.severity]
+        no-blanks-blockquote = 'warn'
+        "#;
+
+        let parsed = parse_toml_config(config_str).unwrap();
+        assert_eq!(
+            RuleSeverity::Warning,
+            *parsed.linters.severity.get("no-blanks-blockquote").unwrap()
+        );
+    }
+
+    #[test]
+    fn test_parse_toml_config_with_invalid_rules() {
+        let config_str = r#"
+        [linters.severity]
+        heading-style = 'err'
+        some-invalid-rule = 'warn'
+
+        [linters.settings.heading-style]
+        style = 'atx'
+        "#;
+
+        let parsed = parse_toml_config(config_str).unwrap();
+        assert_eq!(
+            RuleSeverity::Error,
+            *parsed.linters.severity.get("heading-increment").unwrap()
+        );
+        assert_eq!(
+            RuleSeverity::Error,
+            *parsed.linters.severity.get("heading-style").unwrap()
+        );
+        assert_eq!(None, parsed.linters.severity.get("some-invalid-rule"));
+    }
+
+    #[test]
+    fn test_parse_md004_ul_style_config() {
+        let config_str = r#"
+        [linters.severity]
+        ul-style = 'err'
+
+        [linters.settings.ul-style]
+        style = 'asterisk'
+        "#;
+
+        let parsed = parse_toml_config(config_str).unwrap();
+        assert_eq!(
+            RuleSeverity::Error,
+            *parsed.linters.severity.get("ul-style").unwrap()
+        );
+        assert_eq!(
+            UlStyle::Asterisk,
+            parsed.linters.settings.ul_style.style
+        );
+    }
+
+    #[test]
+    fn test_parse_md004_sublist_style_config() {
+        let config_str = r#"
+        [linters.severity]
+        ul-style = 'warn'
+
+        [linters.settings.ul-style]
+        style = 'sublist'
+        "#;
+
+        let parsed = parse_toml_config(config_str).unwrap();
+        assert_eq!(
+            RuleSeverity::Warning,
+            *parsed.linters.severity.get("ul-style").unwrap()
+        );
+        assert_eq!(
+            UlStyle::Sublist,
+            parsed.linters.settings.ul_style.style
+        );
+    }
+
+    #[test]
+    fn test_parse_md007_ul_indent_config() {
+        let config_str = r#"
+        [linters.severity]
+        ul-indent = 'err'
+
+        [linters.settings.ul-indent]
+        indent = 4
+        start_indent = 3
+        start_indented = true
+        "#;
+
+        let parsed = parse_toml_config(config_str).unwrap();
+        assert_eq!(
+            RuleSeverity::Error,
+            *parsed.linters.severity.get("ul-indent").unwrap()
+        );
+        assert_eq!(4, parsed.linters.settings.ul_indent.indent);
+        assert_eq!(3, parsed.linters.settings.ul_indent.start_indent);
+        assert!(parsed.linters.settings.ul_indent.start_indented);
+    }
+
+    #[test]
+    fn test_parse_md007_default_values() {
+        let config_str = r#"
+        [linters.severity]
+        ul-indent = 'warn'
+        "#;
+
+        let parsed = parse_toml_config(config_str).unwrap();
+        assert_eq!(
+            RuleSeverity::Warning,
+            *parsed.linters.severity.get("ul-indent").unwrap()
+        );
+        // Test default values
+        assert_eq!(2, parsed.linters.settings.ul_indent.indent);
+        assert_eq!(2, parsed.linters.settings.ul_indent.start_indent);
+        assert!(!parsed.linters.settings.ul_indent.start_indented);
+    }
+
+    #[test]
+    fn test_config_from_env_fallback_to_local() {
+        // Create a local config in a temp directory
+        let temp_dir = tempfile::tempdir().unwrap();
+        let config_path = temp_dir.path().join("quickmark.toml");
+        let config_content = r#"
+        [linters.severity]
+        heading-increment = 'err'
+        heading-style = 'off'
+        "#;
+
+        std::fs::write(&config_path, config_content).unwrap();
+
+        // Load config - should fall back to checking the provided path
+        let config = config_from_env_path_or_default(temp_dir.path()).unwrap();
+
+        assert_eq!(
+            RuleSeverity::Error,
+            *config.linters.severity.get("heading-increment").unwrap()
+        );
+        assert_eq!(
+            RuleSeverity::Off,
+            *config.linters.severity.get("heading-style").unwrap()
+        );
+    }
+
+    #[test]
+    fn test_config_from_env_default_when_no_config() {
+        let dummy_path = Path::new("/tmp");
+        let config = config_from_env_path_or_default(dummy_path).unwrap();
+
+        // Should use default configuration
+        assert_eq!(
+            RuleSeverity::Error,
+            *config.linters.severity.get("heading-increment").unwrap()
+        );
+        assert_eq!(
+            RuleSeverity::Error,
+            *config.linters.severity.get("heading-style").unwrap()
         );
     }
 }
