+++ crates/quickmark_server/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0.86"
-quickmark_config = { path = "../quickmark_config" }
 quickmark_linter = { path = "../quickmark_linter" }
 tower-lsp = "0.20.0"
 tokio = { version = "1.0", features = ["full"] }
